GRF-109 improve code hierarchy for Category edit form's attributes factory
